DOP-5071 add a comment and log if function
