+++ populate-metadata/src/assertEnvVars.ts
@@ -14,6 +14,7 @@ export const getEnvVars = (): EnvVars => {
   const environmentVariables = assertEnvVars({
     ATLAS_CLUSTER0_URI: `mongodb+srv://${process.env.MONGO_ATLAS_USERNAME}:${process.env.MONGO_ATLAS_PASSWORD}@${process.env.MONGO_ATLAS_CLUSTER0_HOST}/?retryWrites=true&w=majority`,
     ATLAS_SEARCH_URI: `mongodb+srv://${process.env.MONGO_ATLAS_USERNAME}:${process.env.MONGO_ATLAS_PASSWORD}@${process.env.MONGO_ATLAS_SEARCH_HOST}/?retryWrites=true&w=majority`,
+    // Set to "pool" as a teamwide environment variable and "pool_test" only in the "snooty-mongodb" site
     POOL_DB_NAME: `${process.env.MONGO_ATLAS_POOL_DB_NAME}`,
     REPOS_BRANCHES_COLLECTION: `${process.env.REPOS_BRANCHES_COLLECTION}`,
     DOCSETS_COLLECTION: `${process.env.DOCSETS_COLLECTION}`,
